+++ src/lib/components/data-table/DataTableEvent.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
 	import { TANDAS_NAMES } from '$lib/consts';
-	import DialogToAddPayments from '../DialogToAddPayments.svelte';
+	import DialogToAddPayments from '../SheetToAddPayments.svelte';
 	import { Ticket, CreditCard } from 'lucide-svelte';
 	import { Badge } from '$lib/components/ui/badge';
 	import { Input } from '$lib/components/ui/input';
 	import { page } from '$app/stores';
 	import { formatDateToChile } from '$lib';
-	import DialogToUpdatePayments from '../DialogToUpdatePayments.svelte';
+	import DialogToUpdatePayments from '../SheetToUpdatePayments.svelte';
 	import { formatPriceToCLP } from '$lib';
 	import { Label } from '$lib/components/ui/label/index.js';
 	import { Switch } from '$lib/components/ui/switch/index.js';
@@ -27,6 +27,7 @@
 		customer_phone: string;
 		price: number;
 		ticketAmount: number;
+		ticketValidated: number;
 		ticketsType: string;
 		payment_status: string;
 		buys: {
@@ -58,18 +59,17 @@
 
 	// Reactive statement to filter payments based on search term and switch state
 	$: filteredPayments = Payments.filter((payment) => {
-    const matchesSearchTerm = (
-      payment.customer_name?.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      payment.customer_email?.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      payment.customer_phone?.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      payment.payment_status?.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      payment.rut?.toLowerCase().includes(searchTerm.toLowerCase())
-    );
+		const matchesSearchTerm =
+			payment.customer_name?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+			payment.customer_email?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+			payment.customer_phone?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+			payment.payment_status?.toLowerCase().includes(searchTerm.toLowerCase()) ||
+			payment.rut?.toLowerCase().includes(searchTerm.toLowerCase());
 
-    const matchesSuccessFilter = $showOnlySuccess || payment.payment_status === 'success';
+		const matchesSuccessFilter = $showOnlySuccess || payment.payment_status === 'success';
 
-    return matchesSearchTerm && matchesSuccessFilter;
-  });
+		return matchesSearchTerm && matchesSuccessFilter;
+	});
 </script>
 
 <!-- HEADER -->
@@ -95,6 +95,9 @@
 					class="p-6 w-full hover:bg-muted flex justify-between"
 				>
 					<div class="flex flex-col items-start">
+						<span class="text-xs text-primary uppercase">
+							{payment.client_id ? payment.client_id : '' }
+						</span>
 						<span class=" md:text-xl text-left">
 							{payment.customer_name}
 						</span>
@@ -110,21 +113,21 @@
 						<span class="text-sm">
 							{payment.customer_phone}
 						</span>
+						
 					</div>
 					<div class="flex flex-col items-end gap-4">
 						<Badge
-							class={`${
-								payment.payment_status === 'success'
-									? 'bg-green-400 hover:bg-green-500'
-									: 'bg-red-400 hover:bg-red-500'
-							}`}
+							class={`${payment.payment_status === 'success' ? 'bg-yellow-400' : 'bg-red-400 hover:bg-red-500'}`}
 						>
 							{payment.payment_status}
 						</Badge>
+						{#if payment.ticketAmount === payment.ticketValidated}
+							<Badge class="bg-green-400 hover:bg-green-500">validado</Badge>
+						{/if}
 						<div class="text-right text-xs md:text-base">
 							<span class="flex gap-2 items-center">
 								<Ticket class="h-4 md:h-10" />
-								{payment.ticketAmount}
+								{payment.ticketValidated}/{payment.ticketAmount}
 								{$page.data.eventFromSanityStudio.sell_type === 'batch' ? '' : payment.ticketsType}
 							</span>
 							<span class="flex gap-2 items-center">