+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fix
+
+- No edit payment when is register o rejected.
+
+### Add
+
+- Validate tickets
+
 ## [0.0.10] - 2024-10-28
 
 ### Add